Enhance jwt_decode_handler to handle transition.
As part of ARCH-261, the ecommerce service and ecom-workers will be
transitioning to use the LMS as the sole issuer of JWT tokens. This
update temporarily make the jwt_decode_handler work with several
steps of the final rollout. Ultimately, once the jwt_decode_handler
is only being used to delegate the the edx-drf-extensions version of
the jwt_decode_handler, this custom one will be retired.  See ARCH-261
for the detailed plan.

ARCH-274